[OSX] Build scripts update used for 2009.4.0-RC3 build by Skip Gaede

+++ mac/ExternalPrograms/scripts/enblend.sh
@@ -17,7 +17,9 @@
 #  ppc64ONLYARG="-mcpu=G5 -mtune=G5 -ftree-vectorize" \
 #  OTHERARGs="";
 
-
+# -------------------------------
+# 20091206.0 sg Script tested and used to build 2009.4.0-RC3
+# -------------------------------
 
 # init
 
@@ -48,24 +50,28 @@ do
    TARGET=$i386TARGET
    MACSDKDIR=$i386MACSDKDIR
    ARCHARGs="$i386ONLYARG"
+   OSVERSION="$i386OSVERSION"
    CC=$i386CC
    CXX=$i386CXX
  elif [ $ARCH = "ppc" -o $ARCH = "ppc750" -o $ARCH = "ppc7400" ] ; then
    TARGET=$ppcTARGET
    MACSDKDIR=$ppcMACSDKDIR
    ARCHARGs="$ppcONLYARG"
+   OSVERSION="$ppcOSVERSION"
    CC=$ppcCC
    CXX=$ppcCXX
  elif [ $ARCH = "ppc64" -o $ARCH = "ppc970" ] ; then
    TARGET=$ppc64TARGET
    MACSDKDIR=$ppc64MACSDKDIR
    ARCHARGs="$ppc64ONLYARG"
+   OSVERSION="$ppc64OSVERSION"
    CC=$ppc64CC
    CXX=$ppc64CXX
  elif [ $ARCH = "x86_64" ] ; then
    TARGET=$x64TARGET
    MACSDKDIR=$x64MACSDKDIR
    ARCHARGs="$x64ONLYARG"
+   OSVERSION="$x64OSVERSION"
    CC=$x64CC
    CXX=$x64CXX
  fi
@@ -75,7 +81,7 @@ do
    CFLAGS="-isysroot $MACSDKDIR -arch $ARCH $ARCHARGs $OTHERARGs -dead_strip" \
    CXXFLAGS="-isysroot $MACSDKDIR -arch $ARCH $ARCHARGs $OTHERARGs -dead_strip" \
    CPPFLAGS="-I$REPOSITORYDIR/include -I$REPOSITORYDIR/include/OpenEXR -I/usr/include" \
-   LDFLAGS="-L$REPOSITORYDIR/lib -L/usr/lib -dead_strip" \
+   LDFLAGS="-L$REPOSITORYDIR/lib -L/usr/lib -mmacosx-version-min=$OSVERSION -dead_strip" \
    NEXT_ROOT="$MACSDKDIR" \
    PKG_CONFIG_PATH="$REPOSITORYDIR/lib/pkgconfig" \
    ./configure --prefix="$REPOSITORYDIR" --disable-dependency-tracking --enable-image-cache=yes \
@@ -106,16 +112,28 @@ for program in bin/enblend bin/enfuse
 do
 
  if [ $NUMARCH -eq 1 ] ; then
-   mv "$REPOSITORYDIR/arch/$ARCHS/$program" "$REPOSITORYDIR/$program";
-   strip "$REPOSITORYDIR/$program";
-   continue
+   if [ -f $REPOSITORYDIR/arch/$ARCHS/$program ] ; then
+		 echo "Moving arch/$ARCHS/$program to $program"
+  	 mv "$REPOSITORYDIR/arch/$ARCHS/$program" "$REPOSITORYDIR/$program";
+  	 strip "$REPOSITORYDIR/$program";
+  	 continue
+	 else
+		 echo "Program arch/$ARCHS/$program not found. Aborting build";
+		 exit 1;
+	 fi
  fi
 
  LIPOARGs=""
 
  for ARCH in $ARCHS
  do
-   LIPOARGs="$LIPOARGs $REPOSITORYDIR/arch/$ARCH/$program"
+ 	if [ -f $REPOSITORYDIR/arch/$ARCH/$program ] ; then
+		echo "Adding arch/$ARCH/$program to bundle"
+ 		LIPOARGs="$LIPOARGs $REPOSITORYDIR/arch/$ARCH/$program"
+	else
+		echo "File arch/$ARCH/$program was not found. Aborting build";
+		exit 1;
+	fi
  done
 
  lipo $LIPOARGs -create -output "$REPOSITORYDIR/$program";
